Fix : handle cd() access &  clean up
