+++ src/manda/built_in/Makefile
@@ -8,7 +8,8 @@ endif
 NAME := built_in.a
 HEAD := built_in.h
 SRCS := built_in.c \
-		built_in_funct.c \
+		built_in_env.c \
+		built_in_other.c \
 
 LDLIBS := -lft
