+++ src/manda/built_in/built_in.c
@@ -6,7 +6,7 @@
 /*   By: taerakim <taerakim@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 17:00:41 by taerakim          #+#    #+#             */
-/*   Updated: 2024/04/23 16:06:57 by taerakim         ###   ########.fr       */
+/*   Updated: 2024/04/24 16:14:00 by taerakim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ t_bi_type	is_built_in(char *cmdname)
 {
 	int			i;
 	const int	len = ft_strlen(cmdname);
-	char		*functlist[7] = {"echo", "cd", "pwd", "export" \
+	const char	*functlist[7] = {"echo", "cd", "pwd", "export" \
 								, "unset", "env", "exit"};
 
 	i = 0;
@@ -34,8 +34,8 @@ t_bi_type	is_built_in(char *cmdname)
 
 int	execute_built_in(char **cmds, t_env *env, t_bi_type type)
 {
-	t_built_in	built_in_funct[7] = {ft_echo, ft_cd, ft_pwd, ft_export \
-									, ft_unset , ft_env, ft_exit};
+	const t_built_in	built_in_funct[7] = {ft_echo, ft_cd, ft_pwd, ft_export \
+											, ft_unset, ft_env, ft_exit};
 
 	return (built_in_funct[type](cmds, env));
 }
